Both apr and apr-util are required to build passenger. The are not neccessarily installed in the same --prefix (see ./configure) (note, the apu-1-config does not accept the --cppflags option
With my pgollucci@apache.org HTTPD PMC hat on:

Technicallly, this is still wrong, this should be using apxs alone,
but that would have required me to refactor this and I did not have
time at the moment.   For an example of the correct way to do this,
see the apreq (libapreq2) httpd module (c-binding portion).
[apreq-dev@httpd.apache.org]

Note, Eventually, httpd will have an httpd-config similiar to
gnome-config, neon-config, serf-config, etc.... at which point
it should switch to that.