Updated installation documentation per Phusion wish list:
"The instructions about Linux package installation are wrong. We don’t
provide a .deb (although that was the plan at the time the chapter was
written). Instead, Neil Wilson from Brightbox provides an Apt repository
for Ubuntu. The instructions should be updated accordingly. Correct
instructions can be found at the Phusion blog: see one of the Phusion
Passenger announcements."