Do not canonicalize paths anymore.
Symlinks will now no longer be fully resolved. This makes Phusion Passenger
play nicely with Capistrano-style deployments, in which new releases are
deployed to a new directory, while the document root points to a symlink which
refers to the latest release. The problem with canonicalizing paths was that
new deployments are treated as entirely new applications, thus leaving old
processes behind which should be killed.