Improve the STDERR capturing code. Fixes issue #332.
The old code temporarily points STDERR to a temp file, but doesn't take
into account that the app might create a reference to the current STDERR,
i.e. to that temp file object. It also doesn't take into account that apps
might try to reopen STDERR, or that they might access $stderr instead of
STDERR. The new code fixes all of this.