Fix file descriptor passing problems.
When the ApplicationSpawner server spawns a worker process, do not let it
the worker process communicate directly with the process that issued the
spawn command to the ApplicationSpawner server. This is because upon receiving
the owner pipe, the issuer process will send a message to the
ApplicationSpawner server, as part of the new FD negotiation protocol. If the
ApplicationSpawner server has gone back into its message multiplexing main
loop then it'll crash.

Instead, the ApplicationSpawner server will now act as a proxy between the
issuer and the worker process.