Fix more file descriptor passing oddities.
Once you've sent a file descriptor you expect to be able to close it on the
sender's side, even if the other side hasn't read the file descriptor yet.
Not so: on some operating systems (I'm looking at you OS X) this
can cause the receiving side to receive a bad file descriptor.
The post negotiation protocol ensures that we block until the
other side has really received the file descriptor.