Fix "crude" hook calling order of end_blocking_mod_dir to have parity with Apache's mod_dir.
When Passenger's end_blocking_mod_dir hook is executed, all of its "pre" dependencies, in this case mod_dir, are executed immediately. Unfortunately, this means that mod_dir, which is normally a very low priority HOOK_LAST, gets bumped up to HOOK_MIDDLE when mod_passenger is installed (even if it is inactive, and even if the hooks are totally neutered).

This breaks any other existing web applications (such as a mod_python installation) that are relying on being able to execute before mod_dir by having registered as HOOK_MIDDLE. To fix this, the flags for hooks dependent on mod_dir have been updated to match the calling order of the underlying module from Apache.

(As I do not have a similar issue with mod_autoindex I have not adjusted end_blocking_mod_autoindex to HOOK_MIDDLE; however, it should probably be looked at again, keeping this issue in mind.)