Only prepend LIBDIR to the load path if it's truly missing.
Forcing LIBDIR to be the first item on the load path means that if
Phusion Passenger has been installed as a package in the main Ruby
library directory, then 'rack' (if any) in that directory will always
be used regardless of what the Gemfile may ask for.

With this patch LIBDIR is only prepended to the load path if it is not
present anywhere on the existing path, rather than if it isn't the first
thing on the path. This allows any rack gem added to the path by Bundler
to take precedence over one in LIBDIR.

See also https://github.com/FooBarWidget/passenger/pull/26