AIX has a #define of
#define events req_events

as part of the poll interface.  If you dance just right, you can
include poll.h before you define the structures and your structures end
up with req_events instead of events which is weird but it compiles.

But it is usually easier to just avoid it and make your fields not be
named "events".  So I added a prefix to the two fields that you have.
I generally add prefixes to fields.  It makes cscope more effective and
also avoids name collisions.

I don't know if you want to include this change or not.