Fixed a crash occurs when an application fails to spawn, but the HTTP client disconnects before the error page is generated.
Fixes issue #1028. Calls to clientOutputPipe->write() cause an EPIPE error to be detected,
causing the client to be disconnected. However if there are two clientOutputPipe->write()
calls in succession, then the second one fails with a libeio assertion error because the
client has already been disconnected. The solution is to add more client->connected() checks.