Remove and replace Node.js httplib emulation layer.
Httplib emulation turns out to be a dead end. If we half-close the
session socket, then the 'end' event is emitted right after we're done
forwarding the request headers + body, which is too soon for libraries
like SockJS which expect 'end' to mean "client socket fully closed".
On the other hand, fully closing the session socket doesn't work either
because then we don't know when the request body ends. To fix that we'll
have to introduce HTTP-esque features for determining the size of the
request body, like Content-Length and Transfer-Encoding, which is exactly
what we're trying to avoid.

Furthermore, all the monkey patches are a constant up-hill battle with
compatibility, e.g. https://github.com/andrewrk/node-multiparty/issues/68

So now we're going for a radically new approach. We use the real Node.js
net.HttpServer, have Node.js communicate with Phusion Passenger through
HTTP. Hopefully this fixes all issues once and for all.

See also https://github.com/phusion/passenger/issues/1198#issuecomment-44653299