Properly log uncaught exceptions in Ruby threads
We were relying on Thread#abort_on_exception to do that, but the problem
is that it is implemented by forwarding the uncaught exception to the main
thread, which may not expect that particular exception and may not handle
it properly (such as logging it immediately). The exception could be
forwarded to the main thread during any point of the main thread's execution.

Now, we make the thread responsible for logging the error.