ApplicationPool2: ensure that Process and Session do not have directly dependencies on Group and Pool.
This is done by extracting the information they need into two new classes,
BasicProcessInfo and BasicGroupInfo. Process and Session contain const references
to these objects, and these objects are not changed after construction so
accessing them is thread-safe and does not require locking.

This refactoring allows us to make the ApplicationPool2 locks more fine-grained
in the future.