oxt::dynamic_thread_group: destroy entry point function object before removing thread from list
This fixes a nasty race condition. The function object may contain bound objects.
In the old code, the thread is removed from the thread group before the function
object is destroyed (along with the bound objects). Callers that call
dynamic_thread_group::join_all() may be under the assumption that all threads are
gone, but if they're running destructors in the background then this assumption
does not hold.

We now ensure that the function object is destroyed before the thread is removed
from the thread group.