Watchdog: when gracefully shutting down agents, send SIGTERM to agent only, not to process group
The sub-agents are responsible for shutting down application processes. If
the Watchdog sends SIGTERM to the entire process group, then that interferes
with e.g. the Node.js `PhusionPassenger.on('exit')` event, which worked if
you restart the app or detach an application process, but not if you stop
Passenger.