Ruby: do not trap SIGEXIT
SIGEXIT is a special signal that is invoked just prior to program termination.
In RequestHandler#reset_signal_handlers, we obtain the signal handlers for all
trappable signal so that we can revert them in RequestHandler#revert_signal_handlers.
But trapping a non-trapped 'SIGEXIT' returns nil instead of 'DEFAULT' (as is the case
with other signals). As a result (prior to this fix), RequestHandler#revert_signal_handlers
calls `trap('SIGEXIT', nil)`, which causes the Ruby interpreter to invoke #call on nil
during exit. That sets $ERROR_INFO to NoMethodError.

Closes GH-1730.