[Apache] Fixed buffer limit crash on large file upload. Closes GH-1620.
This happened when the core's disk buffering couldn't keep up with the
client datastream for some time (causing 4 GB memory buffer). The patch
also drastically limits per-client buffer memory usage to 130 KB by
throttling the uploader. Not a factor for Nginx because we use the Nginx
input buffer there instead of our own.