Error page look improved, unified and now always contains error_id.
Production mode apps that fail to spawn now also show the error id in
the error page, so that users (or admins) can communicate this value for
the admin to search for in the logfile, if needed.

The look & feel for production and development error pages was wildy
different and especially the development error page was outdated and
featured Passenger too prominently (giving the impression that Passenger
is responsible for the issue, which is almost never the case). The
designs are now in sync.