Create a private keychain on macOS when run as a daemon user.
On macOS, when Passenger is run as a daemon user, such as when integrated into the system Apache and launched via apachectl/launchd, Passenger now creates a private keychain for use with the security update checker. This fixes a permissions problem that was causing confusing logs and the security update checker to fail.