Handle apr and apu deps on macOS Sierra or later
On macOS Sierra apu-1-config and apr-1-config are not available, but the apr headers are still present. Hongli already fixed compilation in commit 5dcd3ea6 this simply omits the dep checks for these tools in the case where Passenger’s Apache module is being built against the system Apache on macOS sierra or later. Also if using a non-builtin Apache on Sierra or later reccomend installing apr and apu from homebrew.