SpawningKit: use begin and end timestamps to track subprocess step durations
As opposed to using a single 'duration' variable. The problem with that
approach was that it was difficult to time the duration of an exec()
operation. Apps had to take note of the time before the exec(), and
then comparing that with the time after the exec(). If they had to do
that anyway then may as well use begin and end timestamps everywhere.

Also fixed a few journey step timing bugs while we're at it.