Jenkins CI C++ tests: when copying SpawningKit error logs to testlogs directory, take into consideration that not all files are readable
When 'rake test:cxx' is run with SUDO, generated files in /tmp are owned by root which are not readable
by a normal user. If SUDO is set then we use sudo to copy. Otherwise, we copy only the files
that we can read; /tmp may have root-owned files generated by a previous test run.