ConfigKit::Store: fix dynamic default values that depend on other values and are cached
If an option has a dynamic default value, and the default value
function depends on the values of other options, then the previous
implementation did not work correctly because it calls the default
value function on the current config store, which is empty. So
now we simulate an updated config store and call the default value
function on that simulated config store.