Use less fork() and more asyncFork()
On macOS 10.11 El Capitan, we have confirmed
a case in which the child process can get stuck indefinitely with 0% CPU.
If we we create a thread which performs memory allocation, and shortly
after thread creation we fork, then the child process gets stuck because
one of its pthread_atfork() handlers tries to allocate memory, which tries
to grab a lock which was already locked.

This means that on macOS we pretty much should never use regular fork() at
all in a multithreaded environment.

This commit is a first step towards completely getting rid of regular
fork() in favor of asyncFork().