SpawnEnvSetupperMain: do not use fork() directly
On macOS 10.11 El Capitan, we have confirmed that this causes
the child process can get stuck indefinitely with 0% CPU.
The LoggingKit GC thread spawned shortly before we fork,
and that thread grabs some memory allocation mutex while
we are forking. Inside the child, some pthread_atfork()
handler tries to grab that mutex too and deadlocks.

We now use runCommand() which avoids this problem.