Apache module: fix server config setter functions for boolean fields.
The previous code treated such fields the same as a DirConfig Threeway
field, but the ServerConfig uses bools and not Threeways.

While we're at it, fix some related code to cast const char * directly
to bool. Instead, compare with NULL. The semantics are the same but the
code becomes clearer this way.