Fix support for configurations that have two `passenger_base_uri` options in a single virtual host, without corresponding `passenger_app_group_name` and `passenger_app_root` directives
There are two distinct problems.

In the Passenger Core controller, Controller/InternalUtils.cpp
resolveSymlink() has a bug: it was supposed to return
`extractDirName(path to base URI symlink) + "/" + symlink value`,
but it actually returned
`symlink value + "/" + extractDirName(path to base URI symlink)`.
This resulted in a wrong default `passenger_app_root` value to be
calculated.

In the Nginx module, the code that calculates a default
`passenger_app_group_name` value returns
`extractDirName(path to base URI symlink)` but this is
not correct: it should also resolve the symlink. Consider
this example:

    server {
        root /app/public;
        ...
        passenger_base_uri /foo;
        passenger_base_uri /bar;
    }

With /app/public/foo pointing to /foo/public,
and /app/public/bar pointing to /bar/public.

The calculation code thinks, for both requests /foo and /bar,
that the app root is /app/public.

By changing the algorithm to something like this...

    extractDirName(resolveSymlink(path to base URI symlink))

...we solve this problem.

Closes GH-2043.