Fix the fact that log file monitoring keeps appending old data to the internal data structures
The previous behavior makes it seem as if the monitored log file's
content is constantly being appended to its own file. For example
if the log file contained two lines, A and B, then the user may first see:

 A
 B

Then later the user sees:

 A
 B
 A
 B

And then:

 A
 B
 A
 B
 A
 B

And so forth. This commit fixes the bug by replacing the internal
storage with the new `tail -n` output instead of appending the output.
While we're at it, this commit introduces a few micro-optimizations,
such as zero-copy line splitting and reduced hash table lookups.