Debian packaging tests: wait until exactly one Passenger instance exists
Apache restarts Passenger during Apache startup, so there are temporarily
two Passenger instances. This will cause commands like 'passenger-status'
to fail. So we wait until the first instance has exited and the second
instance has fully initialized.

We use `passenger-config list-instances` to list all Passenger instances,
but this command crashes if the Core hasn't started yet (it hasn't written
the PID file yet), so we modify it to not crash when the PID file does
not exist.