SpawningKit: sanity-check Unix domain socket addresses reported by the app
If apps can report arbitrary Unix domain socket filenames then they
can make Passenger connect to (and forward traffic to) arbitrary sockets.
That probably isn't problematic but we don't want to allow it just to be
safe.

A more important issue is this: if any of the parent directories of the
socket is writable by a normal user (Joe) that is not the app's user (Jane),
then Joe can swap that directory with something else, with contents
controlled by Joe. That way, Joe can cause Passenger to connect to (and
forward Jane's traffic to) a process that does not actually belong to Jane.

We mitigate these issues with extra permission checks, and by insisting
that any sockets must be created inside the instance directory's app.s
subdirectory.