SpawningKit: ensure safe reading of files in the work dir after its finalization
After finalization of the work dir, it is owned by the app's user and group.
This allows the app to arbitrarily modify the work dir to insert symlinks.
In order to prevent the SpawningKit code (which may run as root) from reading
arbitrary files through symlink attacks, we use the safeReadFile() function.
This function also mitigates various DoS attacks (see its comments).