Fix forwarding chunked request bodies applications
When request body buffering is disabled in the Core Controller (which
is the case e.g. when using Apache with 'PassengerUploadBuffer off'),
the Core Controller ends up forwarding dechunked body data to the
application, even though it preserves the 'Transfer-Encoding: chunked'
header. The body is automatically dechunked by ServerKit::HttpServer,
so this commit adds a feature to ServerKit::HttpServer to allow disabling
auto-dechunking on a per-request basis. The Core Controller can then
enable auto-dechunking when body buffering is enabled, and disable
auto-dechunking when body buffering is disabled.

There were also a bunch of typos in the Ruby wrapper's request handler:
the code checked for the 'TRANSFER_ENCODING' header, but should actually
check for 'HTTP_TRANSFER_ENCODING'. 'TRANSFER_ENCODING' (without the
'HTTP_' prefix) is the correct name for the *response* object, but not
for the request object. As a result, the Ruby wrapper's request handler
did not properly detect requests with chunked bodies, which broke its
connection keep-alive code and thus also broke subsequent header parsing,
resulting in erroneous "HTTP header size exceeded maximum" errors.
This commit fixes these typos.

Closes #2102.