Get rid of MACOSX_DEPLOYMENT_TARGET setter in Rakefile
It was introduced in commit e48d4c97e7 in order to work around a
bug in the High Sierra compilation toolchain. It would
non-deterministically set __MAC_OS_X_VERSION_MIN_REQUIRED to
a value that's less than 10.13.

This bug no longer exists in the compilation toolchain today,
so we're getting rid of the workaround. Getting rid of this
also happens to fix another issue:

MACOSX_DEPLOYMENT_TARGET was set to PlatformInfo.os_version,
but on High Sierra this returns "10.13.6", which results in
the compiled .o files to contain the header
LC_VERSION_MIN_MACOSX=10.13.6. However, the header is supposed
to contain LC_VERSION_MIN_MACOSX=10.3 instead. The existance of
the patch version in there confuses the compiler if we link
that .o object using an invocation of the compiler that doesn't
have MACOSX_DEPLOYMENT_TARGET set.

Example: my Nginx installation's build system does not set
MACOSX_DEPLOYMENT_TARGET. But when linking Nginx I got warnings
like this:

    ld: warning: object file (<...>/buildout/common/libpassenger_common/JsonTools/CBindings.o)
    was built for newer OSX version (10.13.6) than being linked (10.13)