Vagrant box provisioning: lock `rake` gem installation to v12.2.1
The Vagrant dev box provisioning installs the rake gem. The gem release 12.3.0 upgraded the Ruby dependency to version 2.0. Since the box is a 14.04, which ships with Ruby 1.9.3, the dependency is now broken.

There are therefore two solutions: change the box to use Ruby 2.0, or use the previous rake release.

Based on the rake release history, such upgrade is due to reported problems with 1.9.3, however, assuming that the box itself worked fine before, such solution is likely less drastic and potentially problematic than swithing to Ruby 2.0.