Configurable spawn dir for CageFS (GH-2145)
Add Apache

    InstanceRegistryDir /run/passenger/registry

and Nginx

    passenger_spawn_dir /run/passenger/spawn;

configuration options to set spawn directory used during handshake.
Hardcoded /tmp is unacceptable for CloudLinux CageFS. Full support
would require per-user directories for handshake and registry.
Current implementation is a kind of trade-off, it does not allow
real illusion of the only user on the server.