passenger-memory-stats: Allow skipping Apache/Nginx sections (#2307)
Allow skipping the Apache/Nginx sections of the passenger-memory-stats output, through the `--no-apache` and `--no-nginx` commandline options.

The output without any option specified is the same.

When specifying each of the two options, the section is entirely removed. This causes a difference in spacing between sections, as the Apache section doesn't have leading newlines, while the Nginx/Passenger ones do. Example:

```
$ passenger-mem-stats
Version: 6.0.2
Date   : 2020-09-18 12:17:42 +0200
[APACHE SECTION]

[NGINX SECTION]

[PASSENGER SECTION]
$ passenger-mem-stats --no-nginx
Version: 6.0.2
Date   : 2020-09-18 12:17:42 +0200

[NGINX SECTION]

[PASSENGER SECTION]
```

This is arguably a negligible difference. I originally looked into removing it, but it seems that the Table class introduces a leading newline (after the escape sequences) for Nginx/Passenger, while it doesn't do it for Apache, so I thought it's not really worth changing that.

Closes #2306.