Rescue any `EncodingError` while handling native support lib `LoadError`
In 6.0.18 we started to try to encode the error string into ASCII before
checking it for `dlopen`. This was done to try to handle invalid byte
sequences caused by a bug in Ruby that's been fixed in Ruby 3.3.

I've encountered a similar issue to the author and this fix hasn't fixed
it. Instead, I now receive an `Encoding::UndefinedConversionError` which
isn't handled. I've decided to wrap the whole error-handling logic with
a rescue around `EncodingError` but I'm open to using `StandardError`
here - we're trying to handle an exception and be nice to Passenger
users but ultimately we do not want the failure to load this extension
to cause application errors and it currently does.